Rename SecondVoteItem into SecondVoteCard
