+++ package.json
@@ -88,6 +88,7 @@
 		"react-native-expandable-card": "^1.0.3",
 		"react-router": "^5.1.2",
 		"react-router-dom": "^5.1.2",
+		"react-scroll": "^1.8.1",
 		"reactstrap": "^8.9.0",
 		"resolve": "1.15.0",
 		"resolve-url-loader": "3.1.2",
@@ -96,6 +97,7 @@
 		"seedrandom": "^2.4.4",
 		"semver": "6.3.0",
 		"smart-webcomponents-react": "^8.2.7",
+		"smooth-scrollbar": "^8.6.0",
 		"style-loader": "0.23.1",
 		"terser-webpack-plugin": "2.3.8",
 		"tesseract.js": "^2.1.4",
@@ -115,6 +117,8 @@
 		"@types/enzyme": "^3.10.8",
 		"@types/enzyme-adapter-react-16": "^1.0.6",
 		"@types/jest": "^24.9.1",
+		"@types/jquery": "^3.5.5",
+		"@types/react-scroll": "^1.8.2",
 		"babel-eslint": "10.1.0",
 		"babel-jest": "^24.9.0",
 		"babel-loader": "8.1.0",
