+++ src/components/SecondVoteCard.css
@@ -1,12 +1,55 @@
-.secondvote-card {
-  background-color: rgb(48, 48, 48);
+.secondvote-cards {
+  border-right: transparent;
+  scroll-snap-align: start;
 }
 
-.secondvote-header {
+.secondvote-card {
+  background: #2B2C36;
+  margin-top:10px;
+  margin-bottom: 0px;
+  height: 57px;
+}
+
+.secondvote-card-header {
   display: flex;
   justify-content: space-between;
+  align-items: center;
 }
 
-.secondvote-title {
-  color: white
+.secondvote-card-name {
+  color:white;
+  font-family: 'protipo', sans-serif;
+  font-size: 17px;
+}
+
+.secondvote-card-chance {
+  color:white;
+  font-family: 'protipo', sans-serif;
+  font-size: 17px;
+  font-weight: normal;
+}
+
+
+/* Changeable */
+.clear {
+  opacity:1.0;
+}
+
+.faded {
+  opacity:0.3;
+}
+
+.away {
+  border-right: transparent
+}
+
+.touched {
+  border-right: 8px solid #3969F5;
+}
+
+.notlast {
+  background-color: rgb(48, 48, 48);
+}
+.last {
+  background-color: rgb(1, 73, 121);
 }
