--- src/components/SecondVoteCardItem.css
@@ -1,55 +1,0 @@
-.secondvote-carditem-frame {
-  border-right: transparent;
-  scroll-snap-align: start;
-}
-
-.secondvote-carditem {
-  background: #2B2C36;
-  margin-top:10px;
-  margin-bottom: 0px;
-  height: 57px;
-}
-
-.secondvote-carditem-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.secondvote-carditem-name {
-  color:white;
-  font-family: 'protipo', sans-serif;
-  font-size: 17px;
-}
-
-.secondvote-carditem-chance {
-  color:white;
-  font-family: 'protipo', sans-serif;
-  font-size: 17px;
-  font-weight: normal;
-}
-
-
-/* Changeable */
-.clear {
-  opacity:1.0;
-}
-
-.faded {
-  opacity:0.3;
-}
-
-.away {
-  border-right: transparent
-}
-
-.touched {
-  border-right: 5px solid blue;
-}
-
-.notlast {
-  background-color: rgb(48, 48, 48);
-}
-.last {
-  background-color: rgb(1, 73, 121);
-}
